refactor: rename to dashboards
