+++ src/obsidian/dashboards-embedded-plugin.tsx
